Replace `VerificationKeyStore` by mock or real implementation in tests
