+++ mithril-aggregator/src/store/mod.rs
@@ -4,13 +4,9 @@ mod verification_key_store;
 
 pub use epoch_settings_storer::EpochSettingsStorer;
 pub use pending_certificate_store::*;
-pub use verification_key_store::{VerificationKeyStore, VerificationKeyStorer};
+pub use verification_key_store::VerificationKeyStorer;
 
 #[cfg(test)]
 pub use epoch_settings_storer::FakeEpochSettingsStorer;
-#[cfg(test)]
-pub use verification_key_store::test_suite as verification_key_store_test_suite;
-#[cfg(test)]
-pub(crate) use verification_key_store::test_verification_key_storer;
 #[cfg(test)]
 pub use verification_key_store::MockVerificationKeyStorer;
